mobileface is working perfectly

+++ TestLab/FinalTest/ai-cnn/layers/math_operations.h
@@ -10,7 +10,7 @@
 #include <stdint.h>
 #include "../parameters.h"
 
-void div_by_scalar(const fp_t *input_channel, const uint16_t num_channels, const fp_t* denominator, fp_t* output_channel);
-void norm_naive_l2(const fp_t *input_channel, const uint16_t num_channels, const int32_t param,  fp_t* output_channel);
+void div_by_scalar(const fp_t *input_channel, const int32_t num_channels, const fp_t* denominator, fp_t* output_channel);
+void norm_naive_l2(const fp_t *input_channel, const int32_t num_channels, const int32_t param,  fp_t* output_channel);
 
 #endif //TESTLAYERS_MATH_OPERATIONS_H
